+++ EssentialFeed/EssentialFeedTests/Feed Cache/CacheFeedUseCaseTests.swift
@@ -125,45 +125,6 @@ class CacheFeedUseCaseTests: XCTestCase {
 		XCTAssertEqual(receivedError as NSError?, expectedError, file: file, line: line)
 	}
 	
-	class FeedStoreSpy: FeedStore {
-		enum ReceivedMessage: Equatable {
-			case deleteCachedFeed
-			case insert([LocalFeedImage], Date)
-		}
-		
-		private var deletionCompletions = [DeletionCompletion]()
-		
-		private var insertionCompletions = [InsertionCompletion]()
-		
-		private(set) var receivedMessages = [ReceivedMessage]()
-
-		func deleteCachedFeed(completion: @escaping DeletionCompletion) {
-			deletionCompletions.append(completion)
-			receivedMessages.append(.deleteCachedFeed)
-		}
-		
-		func completeDeletion(with error: Error, at index: Int = 0) {
-			deletionCompletions[index](error)
-		}
-		
-		func completeDeletionSuccessfully(at index: Int = 0) {
-			deletionCompletions[index](nil)
-		}
-
-		func insert(_ feed: [LocalFeedImage], timestamp: Date, completion: @escaping InsertionCompletion) {
-			insertionCompletions.append(completion)
-			receivedMessages.append(.insert(feed, timestamp))
-		}
-		
-		func completeInsertion(with error: Error, at index: Int = 0) {
-			insertionCompletions[index](error)
-		}
-		
-		func completeInsertionSuccessfully(at index: Int = 0) {
-			insertionCompletions[index](nil)
-		}
-	}
-	
 	private func uniqueImage() -> FeedImage {
 		return FeedImage(id: UUID(), description: "any", location: "any", url: anyURL())
 	}
