+++ EssentialFeed/EssentialFeedTests/Feed Cache/LoadFeedFromCacheUseCaseTests.swift
@@ -25,43 +25,4 @@ class LoadFeedFromCacheUseCaseTests: XCTestCase {
 		trackForMemoryLeaks(sut, file: file, line: line)
 		return (sut, store)
 	}
-	
-	private class FeedStoreSpy: FeedStore {
-		enum ReceivedMessage: Equatable {
-			case deleteCachedFeed
-			case insert([LocalFeedImage], Date)
-		}
-		
-		private(set) var receivedMessages = [ReceivedMessage]()
-		
-		private var deletionCompletions = [DeletionCompletion]()
-		private var insertionCompletions = [InsertionCompletion]()
-		
-		func deleteCachedFeed(completion: @escaping DeletionCompletion) {
-			deletionCompletions.append(completion)
-			receivedMessages.append(.deleteCachedFeed)
-		}
-		
-		func completeDeletion(with error: Error, at index: Int = 0) {
-			deletionCompletions[index](error)
-		}
-		
-		func completeDeletionSuccessfully(at index: Int = 0) {
-			deletionCompletions[index](nil)
-		}
-		
-		func insert(_ feed: [LocalFeedImage], timestamp: Date, completion: @escaping InsertionCompletion) {
-			insertionCompletions.append(completion)
-			receivedMessages.append(.insert(feed, timestamp))
-		}
-		
-		func completeInsertion(with error: Error, at index: Int = 0) {
-			insertionCompletions[index](error)
-		}
-		
-		func completeInsertionSuccessfully(at index: Int = 0) {
-			insertionCompletions[index](nil)
-		}
-	}
-	
 }
